api-expedientes/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
celery==5.3.4
redis==5.0.1